ARM: dts: imx6sl-tolino-shine2hd: Add backlight boost
Vendor system has a separate checkbox for increasing brightness of the
backlight combined with a big warning that power consumption will
greatly increase. With that checkbox enabled there is a greater range in
which backlight brightness can be adjusted
Add a switch to achieve the same, since the backlight is actually LED,
this seems to be the most appopiate place to add it. Nothing fits
perfectly.

Signed-off-by: Andreas Kemnade <andreas@kemnade.info>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 9912c251fd57912cda1b933fa661f3373f90466e ]