arm64: dts: qcom: sm8450-nagara: Add gpio line names for TLMM
Sony ever so graciously provides GPIO line names in their downstream
kernel (though sometimes they are not 100% accurate and you can judge
that by simply looking at them and with what drivers they are used).

Add these to the PDX223&224 DTSIs to better document the hardware.

Diff between 223 and 224:
< 	gpio-line-names = "NC", /* GPIO_0 */
< 			  "NC",
< 			  "NC",
< 			  "NC",
> 	gpio-line-names = "TELE_SPI_MISO", /* GPIO_0 */
> 			  "TELE_SPI_MOSI",
> 			  "TELE_SPI_CLK",
> 			  "TELE_SPI_CS_N",
< 			  "PM8010_2_RESET_N",
> 			  "NC",
< 			  "NC",
> 			  "UWIDEC_PWR_EN",
< 			  "TOF_RST_N",
> 			  "NC"
< 			  "QLINK1_REQ",
< 			  "QLINK1_EN", /* GPIO_160 */
< 			  "QLINK1_WMSS_RESET_N",
> 			  "NC",
> 			  "NC", /* GPIO_160 */
> 			  "NC",
The tele lens setup is different on 1 IV and 5 IV and power wiring
is different for some lenses, so it makes sense. As for QLINK, no
idea.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Signed-off-by: Bjorn Andersson <andersson@kernel.org>
Link: https://lore.kernel.org/r/20221117141613.19942-1-konrad.dybcio@linaro.org

[ upstream commit: 15245c93d3c8775a8405c0de18704a8c17a41775 ]