dt-bindings: soc: socionext: Add UniPhier ADAMV block
Add devicetree binding schema for the ADAMV block implemented on Socionext
Uniphier SoCs.

The ADAMV block is analog signal amplifier that is a part of the external
video and audio I/O system. This block is implemented on LD11 and LD20,
and this is defined for controlling audio I/O reset only.

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20221213082449.2721-16-hayashi.kunihiko@socionext.com
Signed-off-by: Rob Herring <robh@kernel.org>

[ upstream commit: 0c00d8d315f09ef8639369a7f12e270b7a86bf79 ]