ARM: dts: colibri-imx6ull: improve wake-up with gpio key
The pin GPIO5_IO01 externally pulls down, it is required to sequentially
connect this pin (signal WAKE_MICO#) to +3v3 and then disconnect it to
trigger a wakeup interrupt.
Adding the flag GPIO_PULL_DOWN allows the system to be woken up just
connecting the pin GPIO5_IO01 to +3v3.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 800a4ea17eb86b1f2cfad9fd7a252a3803bc5264 ]